Frontend matchmaking stuff
Minor changes: BasicCanvas will now halt if the user is not logged in, PlayersQueue identifies the roomID with a span. 

Major changes: functions added to index that take roomID from PlayersQueue, connect to match.php through AJAX which performs logic and gives a response when the game is ready. index.php then redirects to BasicCanvas.php?roomID=(roomID) with a GET request, so now all that's left is the Canvas and database stuff.